add dedicated COPY statement planner
