+++ sql/src/main/java/io/crate/planner/statement/CopyStatementPlanner.java
@@ -1,0 +1,245 @@
+/*
+ * Licensed to Crate under one or more contributor license agreements.
+ * See the NOTICE file distributed with this work for additional
+ * information regarding copyright ownership.  Crate licenses this file
+ * to you under the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.  You may
+ * obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+ * implied.  See the License for the specific language governing
+ * permissions and limitations under the License.
+ *
+ * However, if you have executed another commercial license agreement
+ * with Crate these terms will supersede the license and you may use the
+ * software solely pursuant to the terms of the relevant commercial
+ * agreement.
+ */
+
+package io.crate.planner.statement;
+
+import com.carrotsearch.hppc.procedures.ObjectProcedure;
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.Lists;
+import io.crate.analyze.CopyAnalyzedStatement;
+import io.crate.analyze.WhereClause;
+import io.crate.analyze.symbol.InputColumn;
+import io.crate.analyze.symbol.Reference;
+import io.crate.analyze.symbol.Symbol;
+import io.crate.core.collections.TreeMapBuilder;
+import io.crate.metadata.*;
+import io.crate.metadata.doc.DocSysColumns;
+import io.crate.metadata.doc.DocTableInfo;
+import io.crate.operation.aggregation.impl.CountAggregation;
+import io.crate.planner.Plan;
+import io.crate.planner.Planner;
+import io.crate.planner.distribution.DistributionInfo;
+import io.crate.planner.node.dql.CollectAndMerge;
+import io.crate.planner.node.dql.CollectPhase;
+import io.crate.planner.node.dql.FileUriCollectPhase;
+import io.crate.planner.node.dql.MergePhase;
+import io.crate.planner.projection.Projection;
+import io.crate.planner.projection.SourceIndexWriterProjection;
+import io.crate.planner.projection.WriterProjection;
+import org.apache.lucene.util.BytesRef;
+import org.elasticsearch.cluster.ClusterService;
+import org.elasticsearch.cluster.node.DiscoveryNodes;
+import org.elasticsearch.common.inject.Inject;
+import org.elasticsearch.common.inject.Singleton;
+
+import java.util.*;
+import java.util.concurrent.atomic.AtomicInteger;
+
+@Singleton
+public class CopyStatementPlanner {
+
+    private final ClusterService clusterService;
+
+    @Inject
+    public CopyStatementPlanner(ClusterService clusterService) {
+        this.clusterService = clusterService;
+    }
+
+    public Plan planCopyFrom(CopyAnalyzedStatement analysis, Planner.Context context) {
+        /**
+         * copy from has two "modes":
+         *
+         * 1: non-partitioned tables or partitioned tables with partition ident --> import into single es index
+         *    -> collect raw source and import as is
+         *
+         * 2: partitioned table without partition ident
+         *    -> collect document and partition by values
+         *    -> exclude partitioned by columns from document
+         *    -> insert into es index (partition determined by partition by value)
+         */
+
+        DocTableInfo table = analysis.table();
+        int clusteredByPrimaryKeyIdx = table.primaryKey().indexOf(analysis.table().clusteredBy());
+        List<String> partitionedByNames;
+        String partitionIdent = null;
+
+        List<BytesRef> partitionValues;
+        if (analysis.partitionIdent() == null) {
+
+            if (table.isPartitioned()) {
+                partitionedByNames = Lists.newArrayList(
+                        Lists.transform(table.partitionedBy(), ColumnIdent.GET_FQN_NAME_FUNCTION));
+            } else {
+                partitionedByNames = Collections.emptyList();
+            }
+            partitionValues = ImmutableList.of();
+        } else {
+            assert table.isPartitioned() : "table must be partitioned if partitionIdent is set";
+            // partitionIdent is present -> possible to index raw source into concrete es index
+
+            partitionValues = PartitionName.decodeIdent(analysis.partitionIdent());
+            partitionIdent = analysis.partitionIdent();
+            partitionedByNames = Collections.emptyList();
+        }
+
+        SourceIndexWriterProjection sourceIndexWriterProjection = new SourceIndexWriterProjection(
+                table.ident(),
+                partitionIdent,
+                new Reference(table.getReferenceInfo(DocSysColumns.RAW)),
+                table.primaryKey(),
+                table.partitionedBy(),
+                partitionValues,
+                table.clusteredBy(),
+                clusteredByPrimaryKeyIdx,
+                analysis.settings(),
+                null,
+                partitionedByNames.size() >
+                0 ? partitionedByNames.toArray(new String[partitionedByNames.size()]) : null,
+                table.isPartitioned() // autoCreateIndices
+        );
+        List<Projection> projections = Collections.<Projection>singletonList(sourceIndexWriterProjection);
+        partitionedByNames.removeAll(Lists.transform(table.primaryKey(), ColumnIdent.GET_FQN_NAME_FUNCTION));
+        int referencesSize = table.primaryKey().size() + partitionedByNames.size() + 1;
+        referencesSize = clusteredByPrimaryKeyIdx == -1 ? referencesSize + 1 : referencesSize;
+
+        List<Symbol> toCollect = new ArrayList<>(referencesSize);
+        // add primaryKey columns
+        for (ColumnIdent primaryKey : table.primaryKey()) {
+            toCollect.add(new Reference(table.getReferenceInfo(primaryKey)));
+        }
+
+        // add partitioned columns (if not part of primaryKey)
+        for (String partitionedColumn : partitionedByNames) {
+            toCollect.add(
+                    new Reference(table.getReferenceInfo(ColumnIdent.fromPath(partitionedColumn)))
+            );
+        }
+        // add clusteredBy column (if not part of primaryKey)
+        if (clusteredByPrimaryKeyIdx == -1) {
+            toCollect.add(
+                    new Reference(table.getReferenceInfo(table.clusteredBy())));
+        }
+        // finally add _raw or _doc
+        if (table.isPartitioned() && analysis.partitionIdent() == null) {
+            toCollect.add(new Reference(table.getReferenceInfo(DocSysColumns.DOC)));
+        } else {
+            toCollect.add(new Reference(table.getReferenceInfo(DocSysColumns.RAW)));
+        }
+
+        DiscoveryNodes allNodes = clusterService.state().nodes();
+        FileUriCollectPhase collectPhase = new FileUriCollectPhase(
+                context.jobId(),
+                context.nextExecutionPhaseId(),
+                "copyFrom",
+                generateRouting(allNodes, analysis.settings().getAsInt("num_readers", allNodes.getSize())),
+                table.rowGranularity(),
+                analysis.uri(),
+                toCollect,
+                projections,
+                analysis.settings().get("compression", null),
+                analysis.settings().getAsBoolean("shared", null)
+        );
+
+        return new CollectAndMerge(collectPhase, MergePhase.localMerge(
+                context.jobId(),
+                context.nextExecutionPhaseId(),
+                ImmutableList.<Projection>of(CountAggregation.PARTIAL_COUNT_AGGREGATION_PROJECTION),
+                collectPhase.executionNodes().size(),
+                collectPhase.outputTypes()), context.jobId());
+    }
+
+    public Plan planCopyTo(CopyAnalyzedStatement analysis, Planner.Context context) {
+        DocTableInfo tableInfo = analysis.table();
+        WriterProjection projection = new WriterProjection();
+        projection.uri(analysis.uri());
+        projection.isDirectoryUri(analysis.directoryUri());
+        projection.settings(analysis.settings());
+
+        List<Symbol> outputs;
+        String partitionIdent = analysis.partitionIdent();
+
+        if (analysis.selectedColumns() != null && !analysis.selectedColumns().isEmpty()) {
+            outputs = new ArrayList<>(analysis.selectedColumns().size());
+            List<Symbol> columnSymbols = new ArrayList<>(analysis.selectedColumns().size());
+            for (int i = 0; i < analysis.selectedColumns().size(); i++) {
+                outputs.add(DocReferenceConverter.convertIfPossible(analysis.selectedColumns().get(i), analysis.table()));
+                columnSymbols.add(new InputColumn(i, null));
+            }
+            projection.inputs(columnSymbols);
+        } else {
+            Reference sourceRef;
+            if (analysis.table().isPartitioned() && partitionIdent == null) {
+                // table is partitioned, insert partitioned columns into the output
+                sourceRef = new Reference(analysis.table().getReferenceInfo(DocSysColumns.DOC));
+                Map<ColumnIdent, Symbol> overwrites = new HashMap<>();
+                for (ReferenceInfo referenceInfo : analysis.table().partitionedByColumns()) {
+                    overwrites.put(referenceInfo.ident().columnIdent(), new Reference(referenceInfo));
+                }
+                projection.overwrites(overwrites);
+            } else {
+                sourceRef = new Reference(analysis.table().getReferenceInfo(DocSysColumns.RAW));
+            }
+            outputs = ImmutableList.<Symbol>of(sourceRef);
+        }
+
+        WhereClause where;
+        if (partitionIdent == null) {
+            where = WhereClause.MATCH_ALL;
+        } else {
+            String partitionName = PartitionName.indexName(tableInfo.ident(), partitionIdent);
+            where = new WhereClause(null, null, ImmutableList.of(partitionName));
+        }
+        Routing routing = context.allocateRouting(tableInfo, where, null);
+        CollectPhase collectPhase = new CollectPhase(
+                context.jobId(),
+                context.nextExecutionPhaseId(),
+                "collect",
+                routing,
+                tableInfo.rowGranularity(),
+                outputs,
+                ImmutableList.<Projection>of(projection),
+                WhereClause.MATCH_ALL,
+                DistributionInfo.DEFAULT_BROADCAST
+        );
+        MergePhase mergePhase = MergePhase.localMerge(
+                context.jobId(),
+                context.nextExecutionPhaseId(),
+                ImmutableList.<Projection>of(CountAggregation.PARTIAL_COUNT_AGGREGATION_PROJECTION),
+                collectPhase.executionNodes().size(),
+                collectPhase.outputTypes());
+        return new CollectAndMerge(collectPhase, mergePhase, context.jobId());
+    }
+
+    private static Routing generateRouting(DiscoveryNodes allNodes, int maxNodes) {
+        final AtomicInteger counter = new AtomicInteger(maxNodes);
+        final Map<String, Map<String, List<Integer>>> locations = new TreeMap<>();
+        allNodes.dataNodes().keys().forEach(new ObjectProcedure<String>() {
+            @Override
+            public void apply(String value) {
+                if (counter.getAndDecrement() > 0) {
+                    locations.put(value, TreeMapBuilder.<String, List<Integer>>newMapBuilder().map());
+                }
+            }
+        });
+        return new Routing(locations);
+    }
+}
